requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
-e ./andamios-orm  # Local development installation
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest>=7.4.0
pytest-asyncio
pytest-cov
httpx